+++ src/oci/database/models/cloud_exadata_infrastructure_summary.py
@@ -406,7 +406,7 @@ class CloudExadataInfrastructureSummary(object):
     def cluster_placement_group_id(self):
         """
         Gets the cluster_placement_group_id of this CloudExadataInfrastructureSummary.
-        The `OCID`__ of the cluster placement group of the Exadata Infrastructure.
+        The `OCID`__ of the cluster placement group of the Exadata Infrastructure or Db System.
 
         __ https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm
 
@@ -420,7 +420,7 @@ class CloudExadataInfrastructureSummary(object):
     def cluster_placement_group_id(self, cluster_placement_group_id):
         """
         Sets the cluster_placement_group_id of this CloudExadataInfrastructureSummary.
-        The `OCID`__ of the cluster placement group of the Exadata Infrastructure.
+        The `OCID`__ of the cluster placement group of the Exadata Infrastructure or Db System.
 
         __ https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm