+++ src/oci/database/models/exascale_db_storage_vault.py
@@ -674,7 +674,7 @@ class ExascaleDbStorageVault(object):
     def cluster_placement_group_id(self):
         """
         Gets the cluster_placement_group_id of this ExascaleDbStorageVault.
-        The `OCID`__ of the cluster placement group of the Exadata Infrastructure.
+        The `OCID`__ of the cluster placement group of the Exadata Infrastructure or Db System.
 
         __ https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm
 
@@ -688,7 +688,7 @@ class ExascaleDbStorageVault(object):
     def cluster_placement_group_id(self, cluster_placement_group_id):
         """
         Sets the cluster_placement_group_id of this ExascaleDbStorageVault.
-        The `OCID`__ of the cluster placement group of the Exadata Infrastructure.
+        The `OCID`__ of the cluster placement group of the Exadata Infrastructure or Db System.
 
         __ https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm