+++ src/oci/dbmulticloud/oracle_db_azure_vault_client.py
@@ -197,7 +197,7 @@ class OracleDbAzureVaultClient(object):
         resource_path = "/oracleDbAzureVault/{oracleDbAzureVaultId}/actions/changeCompartment"
         method = "POST"
         operation_name = "change_oracle_db_azure_vault_compartment"
-        api_reference_link = ""
+        api_reference_link = "https://docs.oracle.com/iaas/api/#/en/database-multicloud-integrations/20240501/OracleDbAzureVault/ChangeOracleDbAzureVaultCompartment"
 
         # Don't accept unknown kwargs
         expected_kwargs = [
@@ -314,7 +314,7 @@ class OracleDbAzureVaultClient(object):
         resource_path = "/oracleDbAzureVault"
         method = "POST"
         operation_name = "create_oracle_db_azure_vault"
-        api_reference_link = ""
+        api_reference_link = "https://docs.oracle.com/iaas/api/#/en/database-multicloud-integrations/20240501/OracleDbAzureVault/CreateOracleDbAzureVault"
 
         # Don't accept unknown kwargs
         expected_kwargs = [
@@ -416,7 +416,7 @@ class OracleDbAzureVaultClient(object):
         resource_path = "/oracleDbAzureVault/{oracleDbAzureVaultId}"
         method = "DELETE"
         operation_name = "delete_oracle_db_azure_vault"
-        api_reference_link = ""
+        api_reference_link = "https://docs.oracle.com/iaas/api/#/en/database-multicloud-integrations/20240501/OracleDbAzureVault/DeleteOracleDbAzureVault"
 
         # Don't accept unknown kwargs
         expected_kwargs = [
@@ -529,7 +529,7 @@ class OracleDbAzureVaultClient(object):
         resource_path = "/oracleDbAzureVault/{oracleDbAzureVaultId}"
         method = "GET"
         operation_name = "get_oracle_db_azure_vault"
-        api_reference_link = ""
+        api_reference_link = "https://docs.oracle.com/iaas/api/#/en/database-multicloud-integrations/20240501/OracleDbAzureVault/GetOracleDbAzureVault"
 
         # Don't accept unknown kwargs
         expected_kwargs = [
@@ -683,7 +683,7 @@ class OracleDbAzureVaultClient(object):
         resource_path = "/oracleDbAzureVault"
         method = "GET"
         operation_name = "list_oracle_db_azure_vaults"
-        api_reference_link = ""
+        api_reference_link = "https://docs.oracle.com/iaas/api/#/en/database-multicloud-integrations/20240501/OracleDbAzureVault/ListOracleDbAzureVaults"
 
         # Don't accept unknown kwargs
         expected_kwargs = [
@@ -834,7 +834,7 @@ class OracleDbAzureVaultClient(object):
         resource_path = "/oracleDbAzureVault/{oracleDbAzureVaultId}/actions/refresh"
         method = "POST"
         operation_name = "refresh_oracle_db_azure_vault"
-        api_reference_link = ""
+        api_reference_link = "https://docs.oracle.com/iaas/api/#/en/database-multicloud-integrations/20240501/OracleDbAzureVault/RefreshOracleDbAzureVault"
 
         # Don't accept unknown kwargs
         expected_kwargs = [
@@ -952,7 +952,7 @@ class OracleDbAzureVaultClient(object):
         resource_path = "/oracleDbAzureVault/{oracleDbAzureVaultId}"
         method = "PUT"
         operation_name = "update_oracle_db_azure_vault"
-        api_reference_link = ""
+        api_reference_link = "https://docs.oracle.com/iaas/api/#/en/database-multicloud-integrations/20240501/OracleDbAzureVault/UpdateOracleDbAzureVault"
 
         # Don't accept unknown kwargs
         expected_kwargs = [