+++ src/oci/file_storage/models/client_options.py
@@ -133,7 +133,7 @@ class ClientOptions(object):
         """
         **[Required]** Gets the source of this ClientOptions.
         Clients these options should apply to. Must be a either
-        single IPv4/IPv6 address or single IPv4/IPv6 CIDR block.
+        single IPv4 address or single IPv4 CIDR block.
 
         **Note:** Access will also be limited by any applicable VCN
         security rules and the ability to route IP packets to the
@@ -150,7 +150,7 @@ class ClientOptions(object):
         """
         Sets the source of this ClientOptions.
         Clients these options should apply to. Must be a either
-        single IPv4/IPv6 address or single IPv4/IPv6 CIDR block.
+        single IPv4 address or single IPv4 CIDR block.
 
         **Note:** Access will also be limited by any applicable VCN
         security rules and the ability to route IP packets to the